Added Poll Layout
The challenge project's basic layout specifies where to place the
poll: index.tsx. It was added a simple layout that shows a single
poll radomly plus it answers respectively. Additionally, a check
circle is shown when an answer is clicked. Finally, a simple
counter of the total votes was included.

In conclusion, poll's functionalities completed 70%, No styles and
animations added yet.