Added Styled Poll
Poll did not have styled-components implemented. Styled-component
s were added at the top of in each new component created.

Likewise, the animation of the background percentage growing was
inlcuded via css. Pose package  was not necessary.

As a result, this poll looks alike the screenshots shown as a
reference. Functionalities 100% completed.