Added WindowDimension hook
Before the emplementation of window dimension hook, the poll does
not look good in mobile devices. Thanks to it, the poll's answer
choice shows only the text icon depending on the size of the window.

As a result, better UI and UX in mobile devices. The poll will
look good even on tiny screens starting at 290 pixels.